+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 Changes
 ==
+Next Version (1.x.x)
+--
+
+FEATURES:
+* Add option to recursively analyze cookbook's dependencies (defaults to true)
+
 18 Nov 2013 (1.0.1)
 --